+++ crates/rome_formatter/src/ts/import/literal_export_name.rs
@@ -1,8 +1,8 @@
 use crate::{FormatElement, FormatResult, Formatter, ToFormatElement};
-use rslint_parser::ast::JsModuleSource;
+use rslint_parser::ast::JsLiteralExportName;
 
-impl ToFormatElement for JsModuleSource {
+impl ToFormatElement for JsLiteralExportName {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        formatter.format_token(&self.value_token()?)
+        formatter.format_token(&self.value()?)
     }
 }
