+++ crates/rome_formatter/src/ts/import/shorthand_named_import_specifier.rs
@@ -3,6 +3,6 @@ use rslint_parser::ast::JsShorthandNamedImportSpecifier;
 
 impl ToFormatElement for JsShorthandNamedImportSpecifier {
     fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
-        Ok(formatter.format_node(self.local_name()?)?)
+        formatter.format_node(self.local_name()?)
     }
 }
