+++ crates/rome_formatter/tests/specs/js/module/import/import_specifiers.js
@@ -1,0 +1,14 @@
+import { hey } from "hey"
+import { hey } from "hey";
+import {test} from "foo.json" assert { for: "for" }
+import { // some funky comment
+    loooooooooooooooooooong as moreeeeeeloooooooooooooooooooong,
+    loooooooooooooooooooong2 as moreeeeeeloooooooooooooooooooong2,
+// some other comment
+} from "test";
+import {
+    loooooooooooooooooooong3 as moreeeeeeloooooooooooooooooooong3
+    // some funky comment
+} from "test";
+
+import {loooooooooooooooooooong3,loooooooooooooooooooong5,loooooooooooooooooooong6 } from "boo"
